refactor(src): 이미지 import/export 방식 변경 및 반영
