+++ frontend/src/assets/images/index.ts
@@ -1,0 +1,3 @@
+export { default as logoIcon } from './logo.png';
+export { default as newTemplateIcon } from './newTemplate.png';
+export { default as searchIcon } from './search.png';
